rename scene_manager (typo) to scene
