+++ src/systems/visual/scene.ts
